+++ plone/app/caching/browser/controlpanel.py
@@ -344,6 +344,12 @@ class ControlPanel(BaseView):
             ramCacheCleanupInterval,
         )
 
+        if not enabled and purgingEnabled:
+            IStatusMessage(self.request).addStatusMessage(
+                _(u'Purging ist still enabled while caching is disabled!'),
+                'warning',
+            )
+
         IStatusMessage(self.request).addStatusMessage(
             _(u'Changes saved.'),
             'info',